Refactored, took out Tokenize to its own Lexer class, and rewrote example tests (Still not real tests)
